fix: muda a interface das Seasons para outro arquivo
